Protect fully funded escrow against failed approve tx
