+++ app/components/quests/Quest/components/IndividualApplicantView.tsx
@@ -5,8 +5,8 @@ import { BountyUserType } from "@/prisma/queries/bounty";
 import { MAX_SHORTLIST, smallClickAnimation } from "@/src/constants";
 import { useUserWallet } from "@/src/providers";
 import { useBounty } from "@/src/providers/bountyProvider";
-import { api } from "@/src/utils";
-import { BOUNTY_USER_RELATIONSHIP } from "@/types";
+import { api, updateList } from "@/src/utils";
+import { BOUNTY_USER_RELATIONSHIP, BountyState } from "@/types";
 import { motion } from "framer-motion";
 import { X } from "lucide-react";
 import Image from "next/image";
@@ -15,6 +15,9 @@ import { EApplicantsView } from "./ApplicantsView";
 import CheckpointView from "./CheckpointView";
 import { QuestActionView } from "./QuestActions";
 import DepositCTAModal from "./DepositCTAModal";
+import { useReferral } from "@/src/providers/referralProvider";
+import { PublicKey } from "@solana/web3.js";
+import { addSubmitterFFA } from "@/escrow/adapters";
 
 interface Props {
   selectedSubmitter: BountyUserType;
@@ -39,9 +42,10 @@ const IndividualApplicantView: FC<Props> = ({
   setIsAwaitingResponse,
   depositAmount,
 }) => {
-  const { currentUser } = useUserWallet();
+  const { currentUser, currentWallet, program, provider } = useUserWallet();
   const { currentBounty, setCurrentBounty } = useBounty();
   const { mutateAsync: updateBounty } = api.bountyUsers.update.useMutation();
+  const { getRemainingAccounts, getSubmitterReferrer } = useReferral();
   const { data: quote } = api.quote.getQuoteByBountyAndUser.useQuery(
     {
       bountyId: currentBounty.id,
@@ -56,6 +60,64 @@ const IndividualApplicantView: FC<Props> = ({
 
   const [showModal, setShowModal] = useState(false);
   const [showFundModal, setShowFundModal] = useState(false);
+
+  const handleApproveForQuest = async () => {
+    if (!currentBounty || !selectedSubmitter) return;
+    const toastId = toast.loading("Approving Your Lancer...");
+
+    try {
+      const submitterWallet = new PublicKey(selectedSubmitter.publicKey);
+      const remainingAccounts = await getRemainingAccounts(
+        submitterWallet,
+        new PublicKey(currentBounty?.escrow.mint.publicKey)
+      );
+      const refferer = await getSubmitterReferrer(
+        submitterWallet,
+        new PublicKey(currentBounty?.escrow.mint.publicKey)
+      );
+      console.log("refferer", refferer.toBase58());
+      const signature = await addSubmitterFFA(
+        submitterWallet,
+        currentBounty?.escrow,
+        currentWallet,
+        refferer,
+        remainingAccounts,
+        program,
+        provider
+      );
+      const newRelations = updateList(
+        selectedSubmitter.userid === currentUser.id
+          ? currentBounty?.currentUserRelationsList
+          : [],
+        [BOUNTY_USER_RELATIONSHIP.RequestedLancer],
+        [BOUNTY_USER_RELATIONSHIP.ApprovedSubmitter]
+      );
+      const updatedBounty = await updateBounty({
+        bountyId: currentBounty?.id,
+        userId: selectedSubmitter.userid,
+        currentUserId: currentUser.id,
+        relations: newRelations,
+        state: BountyState.IN_PROGRESS,
+        publicKey: selectedSubmitter.publicKey,
+        escrowId: currentBounty?.escrowid,
+        signature,
+        label: "add-approved-submitter",
+      });
+
+      setCurrentBounty(updatedBounty);
+      toast.success("Successfully approved submitter", { id: toastId });
+    } catch (error) {
+      if (
+        (error.message as string).includes(
+          "Wallet is registered to another user"
+        )
+      ) {
+        toast.error("Wallet is registered to another user", { id: toastId });
+      } else {
+        toast.error("Error approving submitter", { id: toastId });
+      }
+    }
+  };
 
   const handleReject = async () => {
     if (!currentBounty || !selectedSubmitter) return;
@@ -288,7 +350,18 @@ const IndividualApplicantView: FC<Props> = ({
                     {...smallClickAnimation}
                     className="bg-success h-9 w-fit px-4 py-2
                     title-text rounded-md text-white disabled:cursor-not-allowed disabled:opacity-80"
-                    onClick={() => setShowModal(true)}
+                    onClick={() => {
+                      if (
+                        Number(quote.price) -
+                          Number(currentBounty.escrow.amount) >
+                        0
+                      ) {
+                        setShowModal(true);
+                      } else {
+                        // run approval function anyways
+                        handleApproveForQuest();
+                      }
+                    }}
                     disabled={isLoading || isAwaitingResponse}
                   >
                     Select for the Quest
@@ -302,7 +375,7 @@ const IndividualApplicantView: FC<Props> = ({
               prompt="Now that you have selected a Lancer for your Quest, you will need to deposit the remaining amount of the quote into escrow. This will be released to the submitter once you have approved their work. These funds are fully refundable if the Quest is cancelled or the submitter is unable to complete the Quest."
               setShowModal={setShowModal}
               setShowFundModal={setShowFundModal}
-              amount={Number(quote.price) - depositAmount}
+              amount={Number(quote.price) - Number(currentBounty.escrow.amount)}
             />
           )}
         </div>
@@ -311,10 +384,10 @@ const IndividualApplicantView: FC<Props> = ({
         <FundQuestModal
           setShowModal={setShowFundModal}
           setShowFundModal={setShowFundModal}
-          // amount={Number(quote.price) - depositAmount}
-          amount={0.02}
+          amount={Number(quote.price) - Number(currentBounty.escrow.amount)}
+          // amount={0.02}
+          handleApproveForQuest={handleApproveForQuest}
           approving={true}
-          selectedSubmitter={selectedSubmitter}
         />
       )}
     </>
